refactor: extract utils from etna.analysis.eda.plots
